# backend/docs/swagger.yaml
definitions:
...
  domain.CreatePartnerRequest:
    properties:
      image_url:
        type: string
      name:
        type: string
      website_url:
        type: string
    required:
    - image_url
    - name
    - website_url
    type: object
  domain.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  domain.MultiplePartnersResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.PartnerResponse'
        type: array
      meta:
        $ref: '#/definitions/domain.Pagination'
    type: object
...
  domain.Pagination:
    properties:
      limit:
        type: integer
      next_page:
        type: integer
      offset:
        type: integer
      page:
        type: integer
      previous_page:
        type: integer
      total:
        type: integer
    type: object
  domain.PartnerResponse:
    properties:
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      website_url:
        type: string
    type: object
...
  domain.UpdatePartnerRequest:
    properties:
      image_url:
        type: string
      name:
        type: string
      website_url:
        type: string
    required:
    - image_url
    - name
    - website_url
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Testing Swagger APIs.
  termsOfService: http://swagger.io/terms/
  title: Nigus Bingo Desktop APIs
  version: "1.0"
paths:
...
  /partners:
    get:
      description: Get all partners
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.MultiplePartnersResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get all partners
      tags:
      - Partners
    post:
      consumes:
      - application/json
      parameters:
      - description: Partner details
        in: body
        name: partner
        required: true
        schema:
          $ref: '#/definitions/domain.CreatePartnerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.PartnerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - JWT: []
      summary: Create a partner with Image and description
      tags:
      - Partners
  /partners/{id}:
    delete:
      parameters:
      - description: Partner ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - JWT: []
      summary: Delete a partner by ID
      tags:
      - Partners
    get:
      description: Get a partner by ID
      parameters:
      - description: Partner ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.PartnerResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get a partner by ID
      tags:
      - Partners
    put:
      consumes:
      - application/json
      parameters:
      - description: Partner ID
        in: path
        name: id
        required: true
        type: integer
      - description: Partner details
        in: body
        name: partner
        required: true
        schema:
          $ref: '#/definitions/domain.UpdatePartnerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.PartnerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - JWT: []
      summary: Update a partner with Image and description
      tags:
      - Partners
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"